Add new class - Product
